chore(refactor): Separate unwrapSnapshot
