+++ src/database/firebase_list_factory.spec.ts
@@ -2,8 +2,7 @@ import * as firebase from 'firebase/app';
 import { FirebaseListFactory, FirebaseListObservable, FirebaseObjectFactory, onChildAdded, onChildChanged, onChildRemoved, onChildUpdated, AngularFireDatabase, AngularFireDatabaseModule } from './index';
 import { FirebaseApp, FirebaseAppConfig, AngularFireModule} from '../angularfire2';
 import { TestBed, inject } from '@angular/core/testing';
-import { unwrapSnapshot } from './utils';
-import { Query, AFUnwrappedDataSnapshot } from './interfaces';
+import { Query } from './interfaces';
 import { Subscription, Observable, Subject } from 'rxjs';
 import { COMMON_CONFIG } from '../test-config';
 import { _do } from 'rxjs/operator/do';
@@ -651,38 +650,6 @@ describe('FirebaseListFactory', () => {
       });
     });
 
-
-    describe('unwrapSnapshot', () => {
-      let val = { unwrapped: true };
-      let snapshot = {
-        ref: { key: 'key' },
-        val: () => val
-      };
-
-      it('should return an object value with a $key property', () => {
-        const unwrapped = unwrapSnapshot(snapshot as firebase.database.DataSnapshot);
-        expect(unwrapped.$key).toEqual(snapshot.ref.key);
-      });
-
-      it('should return an object value with a $value property if value is scalar', () => {
-        const existsFn = () => { return true; }
-        const unwrappedValue5 = unwrapSnapshot(Object.assign(snapshot, { val: () => 5, exists: existsFn }) as firebase.database.DataSnapshot);
-        const unwrappedValueFalse = unwrapSnapshot(Object.assign(snapshot, { val: () => false, exists: existsFn }) as firebase.database.DataSnapshot);
-        const unwrappedValueLol = unwrapSnapshot(Object.assign(snapshot, { val: () => 'lol', exists: existsFn }) as firebase.database.DataSnapshot);
-
-        expect(unwrappedValue5.$key).toEqual('key');
-        expect(unwrappedValue5.$value).toEqual(5);
-        expect(unwrappedValue5.$exists()).toEqual(true);
-
-        expect(unwrappedValueFalse.$key).toEqual('key');
-        expect(unwrappedValueFalse.$value).toEqual(false);
-        expect(unwrappedValueFalse.$exists()).toEqual(true);
-
-        expect(unwrappedValueLol.$key).toEqual('key');
-        expect(unwrappedValueLol.$value).toEqual('lol');
-        expect(unwrappedValueLol.$exists()).toEqual(true);
-      });
-    });
 
     it('should emit values in the observable creation zone', (done: any) => {
       Zone.current.fork({
