+++ src/database/utils.ts
@@ -1,5 +1,5 @@
 import * as firebase from 'firebase/app';
-import { AFUnwrappedDataSnapshot, PathReference, DatabaseReference } from './interfaces';
+import { PathReference, DatabaseReference } from './interfaces';
 import { FirebaseApp } from '../app/index';
 import { isAbsoluteUrl, isNil, isString } from '../utils';
 
@@ -23,33 +23,6 @@ export interface CheckUrlRef {
   isUrl: () => any;
   isRef: () => any;
   isQuery?: () => any;
-}
-
-/**
- * Unwraps the data returned in the DataSnapshot. Exposes the DataSnapshot key and exists methods through the $key and $exists properties respectively. If the value is primitive, it is unwrapped using a $value property. The $ properies mean they cannot be saved in the Database as those characters are invalid.
- * @param {DataSnapshot} snapshot - The snapshot to unwrap
- * @return AFUnwrappedDataSnapshot
- * @example
- * unwrapSnapshot(snapshot) => { name: 'David', $key: 'david', $exists: Function }
- */
-export function unwrapSnapshot (snapshot:firebase.database.DataSnapshot): AFUnwrappedDataSnapshot {
-  var unwrapped = !isNil(snapshot.val()) ? snapshot.val() : { $value: null };
-  if ((/string|number|boolean/).test(typeof unwrapped)) {
-    unwrapped = {
-      $value: unwrapped
-    };
-  }
-  Object.defineProperty(unwrapped, '$key', {
-    value: snapshot.ref.key,
-    enumerable: false
-  });
-  Object.defineProperty(unwrapped, '$exists', {
-    value: () => {
-      return snapshot.exists();
-    },
-    enumerable: false
-  });
-  return unwrapped;
 }
 
 export function checkForUrlOrFirebaseRef(urlOrRef: string | firebase.database.Reference | firebase.database.Query, cases: CheckUrlRef): any {
